client: Refactor quic_client to nonblocking module
