+++ client/src/nonblocking/mod.rs
@@ -1,2 +1,3 @@
 pub mod pubsub_client;
+pub mod quic_client;
 pub mod rpc_client;
